+++ module/combat/combat.js
@@ -261,7 +261,7 @@ export class WWNCombat extends foundry.documents.Combat {
     const npcs = this.combatants.filter(c => c.actor.type === "monster");
     npcs.forEach(npc => {
       const weapons = npc.token?.delta?.syntheticActor?.items.filter(i => i.type === "weapon");
-      weapons.forEach(weapon => weapon.update({ "system.counter.value": weapon.system.counter.max }));
+      weapons.forEach(weapon => weapon.update({ "system.counter.value": 0 }));
     });
   }